third_party/WebKit/LayoutTests/fast/css/border-image-style-none-expected.txt

PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is true
PASS successfullyParsed is true

TEST COMPLETE
